% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{add_column}
\alias{add_column}
\title{Add a column to the spec or running widget}
\usage{
add_column(x, col, after = NULL)
}
\arguments{
\item{x}{A WebSpec, htmlwidget, or tabviz_proxy}

\item{col}{A ColumnSpec (e.g. from \code{col_text()}, \code{col_numeric()}, ...)}

\item{after}{Field or id of an existing column to insert after. \code{NULL}
(default) appends to the end. Use \code{"__start__"} to prepend.}
}
\value{
The modified input
}
\description{
Add a column to the spec or running widget
}
