% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{airline_delays}
\alias{airline_delays}
\title{Airline Carrier Delay Performance}
\format{
A data frame with 40 rows and 10 variables:
\describe{
\item{carrier}{Airline carrier name}
\item{month}{Month of observation (Jan-Apr)}
\item{delay_vs_avg}{Delay in minutes relative to industry average}
\item{on_time_pct}{Percentage of flights arriving on time}
\item{satisfaction}{Customer satisfaction score (1-5 scale)}
\item{flights}{Number of flights operated}
\item{delay_lower}{Lower bound of delay estimate}
\item{delay_upper}{Upper bound of delay estimate}
\item{trend}{List column containing 12-month delay trend (sparkline data)}
}
}
\source{
Simulated data inspired by U.S. DOT aviation statistics.
}
\usage{
airline_delays
}
\description{
A dataset containing airline performance metrics including delay times,
on-time percentages, and customer satisfaction scores. Data is simulated
but inspired by DOT aviation statistics.
}
\examples{
data(airline_delays)
head(airline_delays)

}
\keyword{datasets}
