% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serialize.R
\name{build_cell_styles_fast}
\alias{build_cell_styles_fast}
\title{Apply pre-compiled cell-style recipes to one row.}
\usage{
build_cell_styles_fast(row, recipes)
}
\description{
Apply pre-compiled cell-style recipes to one row.
}
\keyword{internal}
