% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{clear_edits}
\alias{clear_edits}
\title{Clear all in-widget cell and label edits}
\usage{
clear_edits(x)
}
\arguments{
\item{x}{A tabviz_proxy}
}
\value{
The proxy (invisibly)
}
\description{
Clear all in-widget cell and label edits
}
