% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serialize.R
\name{extract_marker_style}
\alias{extract_marker_style}
\title{Extract marker style from explicit column mappings}
\usage{
extract_marker_style(row, spec)
}
\arguments{
\item{row}{A single row of data}

\item{spec}{The WebSpec containing marker_*_col mappings}
}
\value{
A list of marker style properties or NULL if none set
}
\description{
Extract marker style from explicit column mappings
}
\keyword{internal}
