% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{filter_rows}
\alias{filter_rows}
\title{Filter rows}
\usage{
filter_rows(x, field, operator = c("eq", "neq", "gt", "lt", "contains"), value)
}
\arguments{
\item{x}{A WebSpec, htmlwidget, or tabviz_proxy}

\item{field}{Field to filter on}

\item{operator}{One of "eq", "neq", "gt", "lt", "contains"}

\item{value}{Filter value}
}
\value{
The modified input
}
\description{
Filter rows
}
