% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{invoke_proxy_method}
\alias{invoke_proxy_method}
\title{Internal: invoke a proxy method}
\usage{
invoke_proxy_method(proxy, method, args)
}
\description{
Sends a \code{tabviz-proxy} custom message to the running widget with
\code{list(id, method, args)}.
}
\keyword{internal}
