% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serialize.R
\name{is_categorized_themes}
\alias{is_categorized_themes}
\title{Serialize a Semantics object into the per-token bundle map}
\usage{
is_categorized_themes(x)
}
\description{
Serialize a Semantics object into the per-token bundle map
}
\keyword{InteractionSpec}
\keyword{Serialize}
\keyword{internal}
