% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nse.R
\name{is_na_scalar}
\alias{is_na_scalar}
\title{Check if value is a scalar NA}
\usage{
is_na_scalar(x)
}
\arguments{
\item{x}{Value to check}
}
\value{
Logical
}
\description{
Check if value is a scalar NA
}
\keyword{internal}
