% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{move_column}
\alias{move_column}
\title{Move a column to a new position}
\usage{
move_column(x, field, to)
}
\arguments{
\item{x}{A WebSpec, htmlwidget, or tabviz_proxy}

\item{field}{Field or id of the column to move}

\item{to}{Integer index (1-based) or field name of another column to
position relative to. Integer moves the column to that index.}
}
\value{
The modified input
}
\description{
Move a column to a new position
}
