% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{move_row}
\alias{move_row}
\title{Move a row to a new position}
\usage{
move_row(x, row_id, to)
}
\arguments{
\item{x}{A WebSpec, htmlwidget, or tabviz_proxy}

\item{row_id}{Row id (the value in the label/id column)}

\item{to}{Integer index (1-based) or row id of another row to position relative to}
}
\value{
The modified input
}
\description{
Move a row to a new position
}
