% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nba_efficiency}
\alias{nba_efficiency}
\title{NBA Player Efficiency Ratings}
\format{
A data frame with 30 rows and 12 variables:
\describe{
\item{player}{Player name}
\item{team}{Team abbreviation}
\item{conference}{NBA conference (East or West)}
\item{position}{Primary position (G, F, or C)}
\item{games}{Games played}
\item{ppg}{Points per game}
\item{per}{Player Efficiency Rating}
\item{per_lower}{Lower bound of PER estimate}
\item{per_upper}{Upper bound of PER estimate}
\item{all_star}{Whether player was an All-Star}
\item{award}{Notable award or honor}
}
}
\source{
Simulated data based on typical NBA statistics.
}
\usage{
nba_efficiency
}
\description{
A fun dataset containing NBA player statistics including Player Efficiency
Rating (PER), points per game, and All-Star/award information.
}
\examples{
data(nba_efficiency)
head(nba_efficiency)

}
\keyword{datasets}
