% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{package_themes}
\alias{package_themes}
\title{Available theme presets, organized by category.}
\usage{
package_themes()
}
\value{
A nested list: category -> list of resolved \link{WebTheme} objects.
}
\description{
Available theme presets, organized by category.
}
