% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-v2.R
\name{package_themes_v2}
\alias{package_themes_v2}
\title{All v2 preset themes}
\usage{
package_themes_v2()
}
\value{
Named list of \link{WebTheme2} objects.
}
\description{
Returns a named list of all v2 theme presets. The v2 mirror of
\code{\link[=package_themes]{package_themes()}}.
}
