% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{remove_column}
\alias{remove_column}
\title{Remove a column from the spec or running widget}
\usage{
remove_column(x, field)
}
\arguments{
\item{x}{A WebSpec, htmlwidget, or tabviz_proxy}

\item{field}{Field or id of the column to remove}
}
\value{
The modified input
}
\description{
Remove a column from the spec or running widget
}
