% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderSplitTabviz}
\alias{renderSplitTabviz}
\title{Shiny render function for a split-table tabviz plot}
\usage{
renderSplitTabviz(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns a split-table widget}

\item{env}{The environment in which to evaluate expr}

\item{quoted}{Is expr a quoted expression?}
}
\value{
A Shiny render function
}
\description{
Shiny render function for a split-table tabviz plot
}
