% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{renderTabviz}
\alias{renderTabviz}
\title{Shiny render function for a tabviz plot}
\usage{
renderTabviz(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{An expression that returns a widget from \code{tabviz()}}

\item{env}{The environment in which to evaluate expr}

\item{quoted}{Is expr a quoted expression?}
}
\value{
A Shiny render function
}
\description{
Shiny render function for a tabviz plot
}
