% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{resize_column}
\alias{resize_column}
\title{Resize a column}
\usage{
resize_column(x, field, width)
}
\arguments{
\item{x}{A WebSpec, htmlwidget, or tabviz_proxy}

\item{field}{Field or id of the column}

\item{width}{Width in pixels (>= 40)}
}
\value{
The modified input
}
\description{
Resize a column
}
