% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nse.R
\name{resolve_quosure_expr}
\alias{resolve_quosure_expr}
\title{Resolve a quosure expression}
\usage{
resolve_quosure_expr(quo, data, param_name)
}
\arguments{
\item{quo}{A quosure}

\item{data}{The data frame}

\item{param_name}{Parameter name for errors}
}
\value{
List with col_name and modified data
}
\description{
Resolve a quosure expression
}
\keyword{internal}
