% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serialize.R
\name{serialize_interaction}
\alias{serialize_interaction}
\title{Serialize a Semantics object into the per-token bundle map}
\usage{
serialize_interaction(interaction)
}
\description{
Serialize a Semantics object into the per-token bundle map
}
\keyword{InteractionSpec}
\keyword{Serialize}
\keyword{internal}
