% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serialize.R
\name{serialize_split_table}
\alias{serialize_split_table}
\title{Serialize a SplitForest to a JSON-ready list}
\usage{
serialize_split_table(split_table, include_forest = TRUE)
}
\arguments{
\item{split_table}{A SplitForest object}

\item{include_forest}{Whether to include forest plot data in each spec}
}
\value{
A nested list suitable for jsonlite::toJSON
}
\description{
Converts a SplitForest object containing multiple WebSpec objects into
a format suitable for the split forest htmlwidget.
}
\keyword{internal}
