% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{set_caption}
\alias{set_caption}
\title{Set the plot caption}
\usage{
set_caption(x, text)
}
\arguments{
\item{x}{A WebSpec, htmlwidget, or tabviz_proxy.}

\item{text}{New title text. \code{NULL} clears the title.}
}
\value{
The modified input
}
\description{
Set the plot caption
}
