% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{set_cell}
\alias{set_cell}
\title{Set a single cell value on the running widget}
\usage{
set_cell(x, row_id, field, value)
}
\arguments{
\item{x}{A tabviz_proxy}

\item{row_id}{Row id (value in the id column)}

\item{field}{Data field name of the cell}

\item{value}{New value}
}
\value{
The proxy (invisibly)
}
\description{
Set a single cell value on the running widget
}
