% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-api.R
\name{set_diverging}
\alias{set_diverging}
\title{Set the diverging data scheme and re-resolve.}
\usage{
set_diverging(theme, scheme)
}
\arguments{
\item{theme}{A \link{WebTheme}.}

\item{scheme}{Named diverging scheme reference.}
}
\value{
The re-resolved \link{WebTheme}.
}
\description{
Set the diverging data scheme and re-resolve.
}
