% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{set_row_label}
\alias{set_row_label}
\title{Set a row's label on the running widget}
\usage{
set_row_label(x, row_id, label)
}
\arguments{
\item{x}{A tabviz_proxy}

\item{row_id}{Row id}

\item{label}{New label text}
}
\value{
The proxy (invisibly)
}
\description{
Set a row's label on the running widget
}
