% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-api.R
\name{set_sequential}
\alias{set_sequential}
\title{Set the sequential data scheme and re-resolve.}
\usage{
set_sequential(theme, scheme)
}
\arguments{
\item{theme}{A \link{WebTheme}.}

\item{scheme}{Named sequential scheme reference.}
}
\value{
The re-resolved \link{WebTheme}.
}
\description{
Set the sequential data scheme and re-resolve.
}
