% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{set_watermark}
\alias{set_watermark}
\title{Set (or clear) the plot watermark}
\usage{
set_watermark(x, text)
}
\arguments{
\item{x}{A WebSpec, htmlwidget, or tabviz_proxy.}

\item{text}{Watermark text; \code{NULL} removes the watermark.}
}
\value{
The modified input
}
\description{
Set (or clear) the plot watermark
}
