% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{sort_rows}
\alias{sort_rows}
\title{Sort rows by a column}
\usage{
sort_rows(x, by, direction = c("asc", "desc", "none"))
}
\arguments{
\item{x}{A WebSpec, htmlwidget, or tabviz_proxy}

\item{by}{Field or id of the column to sort by}

\item{direction}{One of "asc", "desc", or "none" (clears sort)}
}
\value{
The modified input
}
\description{
Sort rows by a column
}
