% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{split_tabviz_proxy}
\alias{split_tabviz_proxy}
\title{Create a split-table tabviz proxy object}
\usage{
split_tabviz_proxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{The widget (output) id}

\item{session}{The Shiny session (default: current reactive domain)}
}
\value{
A \code{split_tabviz_proxy} object
}
\description{
Create a split-table tabviz proxy object
}
