% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{splitTabvizOutput}
\alias{splitTabvizOutput}
\title{Shiny output function for a split-table tabviz plot}
\usage{
splitTabvizOutput(outputId, width = "100\%", height = "600px")
}
\arguments{
\item{outputId}{Output variable name}

\item{width}{Widget width (CSS units)}

\item{height}{Widget height (CSS units)}
}
\value{
A Shiny output element
}
\description{
Shiny output function for a split-table tabviz plot
}
