% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{tabviz_proxy}
\alias{tabviz_proxy}
\title{Create a tabviz proxy object}
\usage{
tabviz_proxy(id, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{id}{The widget (output) id}

\item{session}{The Shiny session (default: current reactive domain)}
}
\value{
A \code{tabviz_proxy} object
}
\description{
Builds a proxy that can be used to update a tabviz plot in a Shiny app
without re-rendering the entire widget. Pass the proxy to any fluent
verb (\code{sort_rows()}, \code{filter_rows()}, \code{add_column()}, \code{set_theme()}, ...)
to dispatch an update message to the running widget.
}
