% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{tabvizOutput}
\alias{tabvizOutput}
\title{Shiny output function for a tabviz plot}
\usage{
tabvizOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{Output variable name}

\item{width}{Widget width (CSS units)}

\item{height}{Widget height (CSS units)}
}
\value{
A Shiny output element
}
\description{
Shiny output function for a tabviz plot
}
