% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{toggle_group}
\alias{toggle_group}
\title{Toggle a group's collapsed state}
\usage{
toggle_group(x, group_id, collapsed = NULL)
}
\arguments{
\item{x}{A WebSpec, htmlwidget, or tabviz_proxy}

\item{group_id}{The group id to toggle}

\item{collapsed}{TRUE to collapse, FALSE to expand, NULL to toggle}
}
\value{
The modified input
}
\description{
Toggle a group's collapsed state
}
