% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nse.R
\name{validate_style_result}
\alias{validate_style_result}
\title{Validate style evaluation result}
\usage{
validate_style_result(result, expected_length, param_name, expr)
}
\arguments{
\item{result}{Evaluation result}

\item{expected_length}{Expected vector length}

\item{param_name}{Parameter name for errors}

\item{expr}{Original expression (for error message)}
}
\description{
Ensures the result of a style expression is valid.
}
\keyword{internal}
