% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-components.R
\name{VizBarEffect}
\alias{VizBarEffect}
\title{VizBarEffect: Specification for a bar effect in viz_bar}
\usage{
VizBarEffect(
  value = character(0),
  label = NA_character_,
  color = NA_character_,
  opacity = NA_real_
)
}
\arguments{
\item{value}{Column name containing the bar value}

\item{label}{Display label for this effect in legends}

\item{color}{Optional color for this bar}

\item{opacity}{Optional opacity (0-1)}
}
\description{
VizBarEffect: Specification for a bar effect in viz_bar
}
