% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-components.R
\name{VizViolinEffect}
\alias{VizViolinEffect}
\title{VizViolinEffect: Specification for a violin effect}
\usage{
VizViolinEffect(
  data = character(0),
  label = NA_character_,
  color = NA_character_,
  fill_opacity = 0.5
)
}
\arguments{
\item{data}{Column name containing array data (required)}

\item{label}{Display label for this effect}

\item{color}{Fill color for the violin}

\item{fill_opacity}{Fill opacity (0-1)}
}
\description{
VizViolinEffect: Specification for a violin effect
}
