% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-v2.R
\name{web_theme_cochrane_v2}
\alias{web_theme_cochrane_v2}
\title{Cochrane theme (v2)}
\usage{
web_theme_cochrane_v2()
}
\value{
A \link{WebTheme2}.
}
\description{
Cochrane teal, Arial, very compact density for systematic reviews.
}
