% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{web_theme_dark}
\alias{web_theme_dark}
\title{Dark theme — Catppuccin-inspired dark mode default.}
\usage{
web_theme_dark()
}
\value{
A \link{WebTheme}.
}
\description{
Dark theme — Catppuccin-inspired dark mode default.
}
