% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{web_theme_default}
\alias{web_theme_default}
\title{Default theme (v2)}
\usage{
web_theme_default()
}
\value{
A \link{WebTheme}.
}
\description{
Cyan-on-white, comfortable density, system fonts. v2 of the package
default; replaces \link{web_theme_default} in PR 10.
}
