% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-v2.R
\name{web_theme_default_v2}
\alias{web_theme_default_v2}
\title{Default theme (v2)}
\usage{
web_theme_default_v2()
}
\value{
A \link{WebTheme2}.
}
\description{
Cyan-on-white, comfortable density, system fonts. v2 of the package
default; replaces \link{web_theme_default} in PR 10.
}
