% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-lotr.R
\name{web_theme_dwarven}
\alias{web_theme_dwarven}
\title{Dwarven theme — hammered bronze brand, warm gold decorative.}
\usage{
web_theme_dwarven()
}
\value{
A \link{WebTheme}.
}
\description{
Dwarven theme — hammered bronze brand, warm gold decorative.
}
