% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-lotr.R
\name{web_theme_elvish}
\alias{web_theme_elvish}
\title{Elvish theme — starlit navy brand, mithril silver decorative.}
\usage{
web_theme_elvish()
}
\value{
A \link{WebTheme}.
}
\description{
Elvish theme — starlit navy brand, mithril silver decorative.
}
