% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-v2.R
\name{web_theme_jama_v2}
\alias{web_theme_jama_v2}
\title{JAMA theme (v2)}
\usage{
web_theme_jama_v2()
}
\value{
A \link{WebTheme2}.
}
\description{
All-black-and-white, ultra-compact density, Arial typography.
}
