% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-v2.R
\name{web_theme_lancet_v2}
\alias{web_theme_lancet_v2}
\title{Lancet theme (v2)}
\usage{
web_theme_lancet_v2()
}
\value{
A \link{WebTheme2}.
}
\description{
Lancet navy primary, gold accent, serif typography.
}
