% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{web_theme_nature}
\alias{web_theme_nature}
\title{Nature theme — teal-deep with amber accent.}
\usage{
web_theme_nature()
}
\value{
A \link{WebTheme}.
}
\description{
Nature theme — teal-deep with amber accent.
}
