% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{web_theme_nejm}
\alias{web_theme_nejm}
\title{NEJM theme — crimson with subtle warm paper and tint headers.}
\usage{
web_theme_nejm()
}
\value{
A \link{WebTheme}.
}
\description{
NEJM theme — crimson with subtle warm paper and tint headers.
}
