% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{web_theme_presentation}
\alias{web_theme_presentation}
\title{Presentation theme (v2)}
\usage{
web_theme_presentation()
}
\value{
A \link{WebTheme}.
}
\description{
Oversized markers, thick lines, large typography. For slides/posters.
}
