% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-theme-v2.R
\name{AnnotationCluster}
\alias{AnnotationCluster}
\title{AnnotationCluster: title/subtitle/caption/footnote bindings.}
\description{
Each is a TextRole bundle that defaults to mirroring \verb{theme@text@*}.
}
