% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-serialize.R
\name{build_cell_styles}
\alias{build_cell_styles}
\title{Build per-cell styles from column styleMapping}
\usage{
build_cell_styles(row, columns)
}
\description{
Build per-cell styles from column styleMapping
}
\keyword{internal}
