% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-nse.R
\name{generate_computed_col_name}
\alias{generate_computed_col_name}
\title{Generate a unique column name for computed style}
\usage{
generate_computed_col_name(param_name)
}
\arguments{
\item{param_name}{Base parameter name}
}
\value{
Unique column name string
}
\description{
Generate a unique column name for computed style
}
\keyword{internal}
