% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{package_themes}
\alias{package_themes}
\title{All v2 preset themes}
\usage{
package_themes()
}
\value{
Named list of \link{WebTheme} objects.
}
\description{
Returns a named list of all v2 theme presets. The default category surface
for the in-widget switcher when \code{enable_themes = "default"}.
}
