% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{package_themes}
\alias{package_themes}
\title{All v2 preset themes}
\usage{
package_themes()
}
\value{
Named list of \link{WebTheme} objects.
}
\description{
Returns a named list of all v2 theme presets. The v2 mirror of
\code{\link[=package_themes]{package_themes()}}.
}
