% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-theme.R
\name{package_themes}
\alias{package_themes}
\title{Get all package themes}
\usage{
package_themes()
}
\value{
A named list of WebTheme objects
}
\description{
Returns a named list of all themes distributed with the package.
This is the default value for \code{enable_themes} in \code{web_interaction()}.
}
\examples{
# Get all available themes
themes <- package_themes()
names(themes)

# Use specific themes in interaction
web_interaction(enable_themes = package_themes()[c("default", "modern")])
}
