% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-api.R
\name{set_mode}
\alias{set_mode}
\title{Switch light/dark mode and re-resolve.}
\usage{
set_mode(theme, mode)
}
\arguments{
\item{theme}{A \link{WebTheme}.}

\item{mode}{\code{"light"} or \code{"dark"}.}
}
\value{
The re-resolved \link{WebTheme}.
}
\description{
Switch light/dark mode and re-resolve.
}
