% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{update_data}
\alias{update_data}
\title{Replace the data frame}
\usage{
update_data(x, data)
}
\arguments{
\item{x}{A WebSpec, htmlwidget, or tabviz_proxy}

\item{data}{A data.frame with the same column names as the existing one}
}
\value{
The modified input
}
\description{
Replace the data frame
}
