% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-components.R
\name{web_interaction}
\alias{web_interaction}
\alias{web_interaction_minimal}
\alias{web_interaction_publication}
\title{Create interaction specification}
\usage{
web_interaction(
  show_filters = FALSE,
  show_legend = TRUE,
  enable_sort = TRUE,
  enable_collapse = TRUE,
  enable_select = TRUE,
  enable_hover = TRUE,
  enable_resize = TRUE,
  enable_export = TRUE
)

web_interaction_minimal()

web_interaction_publication()
}
\arguments{
\item{show_filters}{Show filter panel}

\item{show_legend}{Show legend}

\item{enable_sort}{Enable column sorting}

\item{enable_collapse}{Enable group collapsing}

\item{enable_select}{Enable row selection}

\item{enable_hover}{Enable hover effects}

\item{enable_resize}{Enable column resizing}

\item{enable_export}{Enable download/export button}
}
\value{
An InteractionSpec object
}
\description{
Create interaction specification
}
