% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-design.R
\name{web_theme_bauhaus}
\alias{web_theme_bauhaus}
\title{Bauhaus theme — red brand, blue structural decorative, yellow accent.}
\usage{
web_theme_bauhaus()
}
\value{
A \link{WebTheme}.
}
\description{
Bauhaus theme — red brand, blue structural decorative, yellow accent.
}
