% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{web_theme_bmj}
\alias{web_theme_bmj}
\title{BMJ theme — clean blue with vermillion accent.}
\usage{
web_theme_bmj()
}
\value{
A \link{WebTheme}.
}
\description{
BMJ theme — clean blue with vermillion accent.
}
