% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{web_theme_cochrane}
\alias{web_theme_cochrane}
\title{Cochrane theme (v2)}
\usage{
web_theme_cochrane()
}
\value{
A \link{WebTheme}.
}
\description{
Cochrane teal, Arial, very compact density for systematic reviews.
}
