% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-lotr.R
\name{web_theme_hobbit}
\alias{web_theme_hobbit}
\title{Hobbit theme — warm earthen brand, foliage-green decorative,
tint header band. IM Fell English body + display.}
\usage{
web_theme_hobbit()
}
\value{
A \link{WebTheme}.
}
\description{
Hobbit theme — warm earthen brand, foliage-green decorative,
tint header band. IM Fell English body + display.
}
