% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{web_theme_nature}
\alias{web_theme_nature}
\title{Nature theme (v2)}
\usage{
web_theme_nature()
}
\value{
A \link{WebTheme}.
}
\description{
Nature blue, refined red accent, Helvetica Neue, balanced density.
}
