% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-design.R
\name{web_theme_solarized}
\alias{web_theme_solarized}
\title{Solarized theme (light). Source Sans 3 body.}
\usage{
web_theme_solarized()
}
\value{
A \link{WebTheme}.
}
\description{
Solarized theme (light). Source Sans 3 body.
}
