% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-design.R
\name{web_theme_swiss}
\alias{web_theme_swiss}
\title{Swiss theme - International Typographic Style: red brand + grid texture.}
\usage{
web_theme_swiss()
}
\value{
A \link{WebTheme}.
}
\description{
Compact density + tight type (1.2 scale ratio). Inter (Helvetica-feel).
Grid texture showcases the underlying grid system. Achromatic neutrals.
}
