% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-design.R
\name{web_theme_tonal}
\alias{web_theme_tonal}
\title{Tonal theme - Material You-style: raised shell, soft elevation,
friendly type rhythm.}
\usage{
web_theme_tonal()
}
\value{
A \link{WebTheme}.
}
\description{
Roboto family + ease curves give the characteristic Material You feel.
}
