% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-design.R
\name{web_theme_tonal}
\alias{web_theme_tonal}
\title{Tonal theme (light) — M3-inspired tonal palette.}
\usage{
web_theme_tonal()
}
\value{
A \link{WebTheme}.
}
\description{
Tonal theme (light) — M3-inspired tonal palette.
}
