% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-design.R
\name{web_theme_tonal_dark}
\alias{web_theme_tonal_dark}
\title{Tonal Dark theme - dark polarity sibling of Tonal.}
\usage{
web_theme_tonal_dark()
}
\value{
A \link{WebTheme}.
}
\description{
Tonal Dark theme - dark polarity sibling of Tonal.
}
