% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes-design.R
\name{web_theme_tufte}
\alias{web_theme_tufte}
\title{Tufte theme — minimal greys; compact density. ET Book-inspired
Crimson Pro serif for body.}
\usage{
web_theme_tufte()
}
\value{
A \link{WebTheme}.
}
\description{
Tufte theme — minimal greys; compact density. ET Book-inspired
Crimson Pro serif for body.
}
